# vmmc backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/compliance/tracker": {
      "get": {
        "operationId": "ComplianceController_getTracker",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Matches against full name or Employee ID.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "required": false,
            "in": "query",
            "description": "Department code (e.g. OPD, RAD, DIET, ADM).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "compliance"
        ]
      }
    },
    "/api/compliance/employees/{id}/dossier": {
      "get": {
        "operationId": "ComplianceController_getDossier",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "compliance"
        ]
      }
    },
    "/api/documents": {
      "post": {
        "operationId": "DocumentsController_upload",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated key so a retried request never creates a duplicate record.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "examDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "cxrResult": {
                    "type": "string",
                    "enum": [
                      "CLEARED",
                      "INFILTRATE"
                    ]
                  },
                  "genexpertResult": {
                    "type": "string",
                    "enum": [
                      "NOT_DETECTED",
                      "DETECTED"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/api/documents/{id}/review": {
      "patch": {
        "operationId": "DocumentsController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
...
  },
  "info": {
    "title": "VMMC TB DOTS API",
    "description": "ITIL 4 Monitoring & Event Management backend for the VMMC TB DOTS health management system. Write endpoints accept an `Idempotency-Key` header so a client retry after a timeout never duplicates a record.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "example": "VMMC-23-0002"
          },
          "password": {
            "type": "string"
          },
          "portal": {
            "type": "string",
            "enum": [
              "staff",
              "admin"
            ]
          }
        },
        "required": [
          "employeeId",
          "password",
          "portal"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "contact": {
            "type": "string",
            "description": "Either an email address or a PH mobile number.",
            "example": "ana.garcia@vmmc.gov.ph"
          }
        },
        "required": [
          "contact"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "contact": {
            "type": "string",
            "description": "Same contact used in the forgot-password request."
          },
          "otp": {
            "type": "string",
            "description": "6-digit code.",
            "example": "123456"
          }
        },
        "required": [
          "contact",
          "otp"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "resetToken": {
            "type": "string",
            "description": "Short-lived token returned by /auth/verify-otp."
          },
          "newPassword": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "resetToken",
          "newPassword"
        ]
      },
...
      "ReviewDocumentDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "approve",
              "reject"
            ]
          },
          "reason": {
            "type": "string",
            "description": "Required when action is \"reject\"."
          },
          "signatureId": {
            "type": "string",
            "description": "Required when action is \"approve\" — the admin's digital_signatures id."
          }
        },
        "required": [
          "action"
        ]
      },
...
    }
  }
}